+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod driver;
 mod types;
 mod registers;
 pub(crate) mod calculate;
+mod validate;