+++ CHANGELOG.md
@@ -14,3 +14,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `set_frequency`, `set_rx_config` methods
 - `RssiSmoothing` enum and `set_rssi_smoothing` method
 - `restart_rx` method
+- `set_rssi_offset` method